requirements.txt
Pillow
torch
numpy
transformers==3.3.0
nltk==3.5
rake_nltk
sentencepiece
fastapi[all]
pytesseract
tinydb